requirements.txt
Django
django-recaptcha==4.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
fontawesomefree
requests
gunicorn
django-storages 
boto3
Pillow
markdown
django-filter
psycopg2-binary==2.9.9